requirements.txt
appdirs==1.4.4
distlib==0.3.1
Django==2.2
django-cors-headers==3.5.0
djangorestframework==3.11.1
psycopg2==2.8.6
filelock==3.0.12
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
asgiref==3.3.1
djangorestframework-simplejwt==4.4.0
mysqlclient==2.0.3
PyJWT==1.7.1
